+++ newsj-util/src/main/java/net/longfalcon/newsj/persistence/MovieInfoDAO.java
@@ -35,7 +35,7 @@ public interface MovieInfoDAO {
 
     MovieInfo findByMovieInfoId(long id);
 
-    MovieInfo findByImdbId(long imdbId);
+    MovieInfo findByImdbId(int imdbId);
 
     List<MovieInfo> findByImdbId(List<Long> imdbIds);
